+++ odbc-api/src/lib.rs
@@ -9,20 +9,24 @@
 mod connection;
 mod cursor;
 mod environment;
-mod into_input;
+mod into_parameter;
 mod parameter_collection;
 mod prepared;
+mod fixed_sized;
 
 pub mod buffers;
 pub mod guide;
 pub mod handles;
+pub mod parameter;
 
 pub use self::{
     connection::Connection,
     cursor::{Cursor, CursorImpl, RowSetBuffer, RowSetCursor},
     environment::Environment,
-    handles::{ColumnDescription, DataType, Error, Input, Nullable},
-    into_input::IntoInput,
+    fixed_sized::Bit,
+    handles::{ColumnDescription, DataType, Error, Nullable},
+    into_parameter::IntoParameter,
+    parameter::Parameter,
     parameter_collection::ParameterCollection,
     prepared::Prepared,
 };