+++ odbc-api/src/parameter_collection/tuple.rs
@@ -1,8 +1,8 @@
 //! Implement `Parameters` trait for tuples consisting of elements implementing `SingleParameter`
 //! trait.
 
-use super::{Input, ParameterCollection};
-use crate::{handles::Statement, Error};
+use super::ParameterCollection;
+use crate::{Parameter, handles::Statement, Error};
 
 macro_rules! impl_bind_input_parameters {
     ($offset:expr, $stmt:ident) => (
@@ -21,7 +21,7 @@ macro_rules! impl_parameters_for_tuple{
         #[allow(unused_parens)]
         #[allow(unused_variables)]
         #[allow(non_snake_case)]
-        unsafe impl<$($t:Input,)*> ParameterCollection for ($($t,)*)
+        unsafe impl<$($t:Parameter,)*> ParameterCollection for ($($t,)*)
         {
             fn parameter_set_size(&self) -> u32 {
                 1