+++ odbcsv/src/main.rs
@@ -1,6 +1,6 @@
 use anyhow::{bail, Error};
 use log::info;
-use odbc_api::{buffers::TextRowSet, Connection, Cursor, Environment, IntoInput};
+use odbc_api::{buffers::TextRowSet, Connection, Cursor, Environment, IntoParameter};
 use std::{
     fs::File,
     io::{stdin, stdout, Read, Write},
@@ -169,7 +169,7 @@ fn query(environment: &Environment, opt: &QueryOpt) -> Result<(), Error> {
     let mut connection = open_connection(&environment, connect_opts)?;
 
     // Convert the input strings into parameters suitable to for use with ODBC.
-    let params: Vec<_> = parameters.iter().map(|param| param.into_input()).collect();
+    let params: Vec<_> = parameters.iter().map(|param| param.into_parameter()).collect();
 
     // Execute the query as a one off, and pass the parameters.
     match connection.execute(&query, params.as_slice())? {